+++ packages/composites/set-token-index/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @chainlink/set-token-index-adapter
 
+## 3.0.50
+
+### Patch Changes
+
+- Updated dependencies [[`7e61390`](https://github.com/smartcontractkit/external-adapters-js/commit/7e613904f77cfe90735b73db07f5a22e651e3810)]:
+  - @chainlink/token-allocation-adapter@1.14.9
+
 ## 3.0.49
 
 ### Patch Changes